split clients
