+++ sdk/timeseriesinsights/Azure.IoT.TimeSeriesInsights/src/QueryClient.cs
@@ -1,0 +1,508 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Azure.Core;
+using Azure.Core.Pipeline;
+
+namespace Azure.IoT.TimeSeriesInsights
+{
+    /// <summary>
+    /// Query client that can be used to perform query operations on Time Series Insights.
+    /// </summary>
+    public class QueryClient
+    {
+        private readonly QueryRestClient _queryRestClient;
+        private readonly ClientDiagnostics _clientDiagnostics;
+
+        /// <summary>
+        /// Initializes a new instance of QueryClient. This constructor should only be used for mocking purposes.
+        /// </summary>
+        protected QueryClient()
+        {
+        }
+
+        internal QueryClient(QueryRestClient queryRestClient, ClientDiagnostics clientDiagnostics)
+        {
+            Argument.AssertNotNull(queryRestClient, nameof(queryRestClient));
+            Argument.AssertNotNull(clientDiagnostics, nameof(clientDiagnostics));
+
+            _queryRestClient = queryRestClient;
+            _clientDiagnostics = clientDiagnostics;
+        }
+
+        /// <summary>
+        /// Retrieve raw events for a given Time Series Id asynchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve raw events for.</param>
+        /// <param name="startTime">Start timestamp of the time range. Events that have this timestamp are included.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded.</param>
+        /// <param name="options">Optional parameters to use when querying for events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual AsyncPageable<QueryResultPage> GetEventsAsync(
+            TimeSeriesId timeSeriesId,
+            DateTimeOffset startTime,
+            DateTimeOffset endTime,
+            QueryEventsRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+            scope.Start();
+
+            try
+            {
+                var searchSpan = new DateTimeRange(startTime, endTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetEvents = new GetEvents(timeSeriesId, searchSpan)
+                };
+
+                BuildEventsRequestOptions(options, queryRequest);
+
+                return QueryInternalAsync(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve raw events for a given Time Series Id synchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve raw events for.</param>
+        /// <param name="startTime">Start timestamp of the time range. Events that have this timestamp are included.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded.</param>
+        /// <param name="options">Optional parameters to use when querying for events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual Pageable<QueryResultPage> GetEvents(
+            TimeSeriesId timeSeriesId,
+            DateTimeOffset startTime,
+            DateTimeOffset endTime,
+            QueryEventsRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+            scope.Start();
+
+            try
+            {
+                var searchSpan = new DateTimeRange(startTime, endTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetEvents = new GetEvents(timeSeriesId, searchSpan)
+                };
+
+                BuildEventsRequestOptions(options, queryRequest);
+
+                return QueryInternal(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve raw events for a given Time Series Id over a specified time interval asynchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve raw events for.</param>
+        /// <param name="timeSpan">The time interval over which to query data.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded. If null is provided, <c>DateTimeOffset.UtcNow</c> is used.</param>
+        /// <param name="options">Optional parameters to use when querying for events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames. Each frame is an array of
+        /// timestamps along with their associated properties.</returns>
+        public virtual AsyncPageable<QueryResultPage> GetEventsAsync(
+            TimeSeriesId timeSeriesId,
+            TimeSpan timeSpan,
+            DateTimeOffset? endTime = null,
+            QueryEventsRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+            scope.Start();
+
+            try
+            {
+                DateTimeOffset rangeEndTime = endTime ?? DateTimeOffset.UtcNow;
+                DateTimeOffset rangeStartTime = rangeEndTime - timeSpan;
+                var searchSpan = new DateTimeRange(rangeStartTime, rangeEndTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetEvents = new GetEvents(timeSeriesId, searchSpan)
+                };
+
+                BuildEventsRequestOptions(options, queryRequest);
+
+                return QueryInternalAsync(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve raw events for a given Time Series Id over a certain time interval synchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve raw events for.</param>
+        /// <param name="timeSpan">The time interval over which to query data.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded. If null is provided, <c>DateTimeOffset.UtcNow</c> is used.</param>
+        /// <param name="options">Optional parameters to use when querying for events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual Pageable<QueryResultPage> GetEvents(
+            TimeSeriesId timeSeriesId,
+            TimeSpan timeSpan,
+            DateTimeOffset? endTime = null,
+            QueryEventsRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+            scope.Start();
+
+            try
+            {
+                DateTimeOffset rangeEndTime = endTime ?? DateTimeOffset.UtcNow;
+                DateTimeOffset rangeStartTime = rangeEndTime - timeSpan;
+                var searchSpan = new DateTimeRange(rangeStartTime, rangeEndTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetEvents = new GetEvents(timeSeriesId, searchSpan)
+                };
+
+                BuildEventsRequestOptions(options, queryRequest);
+
+                return QueryInternal(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve series events for a given Time Series Id asynchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve series events for.</param>
+        /// <param name="startTime">Start timestamp of the time range. Events that have this timestamp are included.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded.</param>
+        /// <param name="options">Optional parameters to use when querying for series events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual AsyncPageable<QueryResultPage> GetSeriesAsync(
+            TimeSeriesId timeSeriesId,
+            DateTimeOffset startTime,
+            DateTimeOffset endTime,
+            QuerySeriesRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetSeries)}");
+            scope.Start();
+
+            try
+            {
+                var searchSpan = new DateTimeRange(startTime, endTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetSeries = new GetSeries(timeSeriesId, searchSpan)
+                };
+
+                BuildSeriesRequestOptions(options, queryRequest);
+
+                return QueryInternalAsync(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve series events for a given Time Series Id synchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve series events for.</param>
+        /// <param name="startTime">Start timestamp of the time range. Events that have this timestamp are included.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded.</param>
+        /// <param name="options">Optional parameters to use when querying for series events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual Pageable<QueryResultPage> GetSeries(
+            TimeSeriesId timeSeriesId,
+            DateTimeOffset startTime,
+            DateTimeOffset endTime,
+            QuerySeriesRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetSeries)}");
+            scope.Start();
+
+            try
+            {
+                var searchSpan = new DateTimeRange(startTime, endTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetSeries = new GetSeries(timeSeriesId, searchSpan)
+                };
+
+                BuildSeriesRequestOptions(options, queryRequest);
+
+                return QueryInternal(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve series events for a given Time Series Id over a specified time interval asynchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve series events for.</param>
+        /// <param name="timeSpan">The time interval over which to query data.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded. If null is provided, <c>DateTimeOffset.UtcNow</c> is used.</param>
+        /// <param name="options">Optional parameters to use when querying for series events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual AsyncPageable<QueryResultPage> GetSeriesAsync(
+            TimeSeriesId timeSeriesId,
+            TimeSpan timeSpan,
+            DateTimeOffset? endTime = null,
+            QuerySeriesRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetSeries)}");
+            scope.Start();
+
+            try
+            {
+                DateTimeOffset rangeEndTime = endTime ?? DateTimeOffset.UtcNow;
+                DateTimeOffset rangeStartTime = rangeEndTime - timeSpan;
+                var searchSpan = new DateTimeRange(rangeStartTime, rangeEndTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetSeries = new GetSeries(timeSeriesId, searchSpan)
+                };
+
+                BuildSeriesRequestOptions(options, queryRequest);
+
+                return QueryInternalAsync(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>
+        /// Retrieve series events for a given Time Series Id over a certain time interval synchronously.
+        /// </summary>
+        /// <param name="timeSeriesId">The Time Series Id to retrieve series events for.</param>
+        /// <param name="timeSpan">The time interval over which to query data.</param>
+        /// <param name="endTime">End timestamp of the time range. Events that match this timestamp are excluded. If null is provided, <c>DateTimeOffset.UtcNow</c> is used.</param>
+        /// <param name="options">Optional parameters to use when querying for series events.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The pageable list <see cref="AsyncPageable{QueryResultPage}"/> of query result frames.</returns>
+        public virtual Pageable<QueryResultPage> GetSeries(
+            TimeSeriesId timeSeriesId,
+            TimeSpan timeSpan,
+            DateTimeOffset? endTime = null,
+            QuerySeriesRequestOptions options = null,
+            CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetSeries)}");
+            scope.Start();
+
+            try
+            {
+                DateTimeOffset rangeEndTime = endTime ?? DateTimeOffset.UtcNow;
+                DateTimeOffset rangeStartTime = rangeEndTime - timeSpan;
+                var searchSpan = new DateTimeRange(rangeStartTime, rangeEndTime);
+                var queryRequest = new QueryRequest
+                {
+                    GetSeries = new GetSeries(timeSeriesId, searchSpan)
+                };
+
+                BuildSeriesRequestOptions(options, queryRequest);
+
+                return QueryInternal(queryRequest, options?.StoreType?.ToString(), cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        private AsyncPageable<QueryResultPage> QueryInternalAsync(
+                    QueryRequest queryRequest,
+                    string storeType, CancellationToken cancellationToken)
+        {
+            async Task<Page<QueryResultPage>> FirstPageFunc(int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+                scope.Start();
+                try
+                {
+                    Response<QueryResultPage> response = await _queryRestClient
+                        .ExecuteAsync(queryRequest, storeType, null, null, cancellationToken)
+                        .ConfigureAwait(false);
+
+                    var frame = new QueryResultPage[]
+                    {
+                            response.Value
+                    };
+
+                    return Page.FromValues(frame, response.Value.ContinuationToken, response.GetRawResponse());
+                }
+                catch (Exception ex)
+                {
+                    scope.Failed(ex);
+                    throw;
+                }
+            }
+
+            async Task<Page<QueryResultPage>> NextPageFunc(string nextLink, int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(GetEvents)}");
+                scope.Start();
+                try
+                {
+                    Response<QueryResultPage> response = await _queryRestClient
+                        .ExecuteAsync(queryRequest, storeType, nextLink, null, cancellationToken)
+                        .ConfigureAwait(false);
+
+                    var frame = new QueryResultPage[]
+                    {
+                            response.Value
+                    };
+
+                    return Page.FromValues(frame, response.Value.ContinuationToken, response.GetRawResponse());
+                }
+                catch (Exception ex)
+                {
+                    scope.Failed(ex);
+                    throw;
+                }
+            }
+
+            return PageableHelpers.CreateAsyncEnumerable(FirstPageFunc, NextPageFunc);
+        }
+
+        private Pageable<QueryResultPage> QueryInternal(
+            QueryRequest queryRequest,
+            string storeType,
+            CancellationToken cancellationToken)
+        {
+            Page<QueryResultPage> FirstPageFunc(int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(QueryInternal)}");
+                scope.Start();
+                try
+                {
+                    Response<QueryResultPage> response = _queryRestClient
+                        .Execute(queryRequest, storeType, null, null, cancellationToken);
+
+                    var frame = new QueryResultPage[]
+                    {
+                            response.Value
+                    };
+
+                    return Page.FromValues(frame, response.Value.ContinuationToken, response.GetRawResponse());
+                }
+                catch (Exception ex)
+                {
+                    scope.Failed(ex);
+                    throw;
+                }
+            }
+
+            Page<QueryResultPage> NextPageFunc(string nextLink, int? pageSizeHint)
+            {
+                using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TimeSeriesInsightsClient)}.{nameof(QueryInternal)}");
+                scope.Start();
+                try
+                {
+                    Response<QueryResultPage> response = _queryRestClient
+                        .Execute(queryRequest, storeType, nextLink, null, cancellationToken);
+
+                    var frame = new QueryResultPage[]
+                    {
+                            response.Value
+                    };
+
+                    return Page.FromValues(frame, response.Value.ContinuationToken, response.GetRawResponse());
+                }
+                catch (Exception ex)
+                {
+                    scope.Failed(ex);
+                    throw;
+                }
+            }
+
+            return PageableHelpers.CreateEnumerable(FirstPageFunc, NextPageFunc);
+        }
+
+        private static void BuildEventsRequestOptions(QueryEventsRequestOptions options, QueryRequest queryRequest)
+        {
+            if (options != null)
+            {
+                if (options.Filter != null)
+                {
+                    queryRequest.GetEvents.Filter = new TimeSeriesExpression(options.Filter);
+                }
+
+                if (options.ProjectedProperties != null)
+                {
+                    foreach (EventProperty projectedProperty in options.ProjectedProperties)
+                    {
+                        queryRequest.GetEvents.ProjectedProperties.Add(projectedProperty);
+                    }
+                }
+
+                queryRequest.GetEvents.Take = options.MaximumNumberOfEvents;
+            }
+        }
+
+        private static void BuildSeriesRequestOptions(QuerySeriesRequestOptions options, QueryRequest queryRequest)
+        {
+            if (options != null)
+            {
+                if (options.Filter != null)
+                {
+                    queryRequest.GetSeries.Filter = new TimeSeriesExpression(options.Filter);
+                }
+
+                if (options.ProjectedVariables != null)
+                {
+                    foreach (string projectedVariable in options.ProjectedVariables)
+                    {
+                        queryRequest.GetSeries.ProjectedVariables.Add(projectedVariable);
+                    }
+                }
+
+                if (options.InlineVariables != null)
+                {
+                    foreach (string inlineVariableKey in options.InlineVariables.Keys)
+                    {
+                        queryRequest.GetSeries.InlineVariables[inlineVariableKey] = options.InlineVariables[inlineVariableKey];
+                    }
+                }
+
+                queryRequest.GetSeries.Take = options.MaximumNumberOfEvents;
+            }
+        }
+    }
+}
