+++ sdk/timeseriesinsights/Azure.IoT.TimeSeriesInsights/tests/SessionRecords/TimeSeriesIdTests/TimeSeriesId_CreateInstanceWith3Keys.json
@@ -8,18 +8,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "97",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "740eb16e45a8acb13310b2f6ef70a52a",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "234cfc8bc0cf22a73d9d6793ac5abaaf",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "put": [
           {
             "timeSeriesId": [
-              "1uotX",
+              "URmuC",
               null,
-              "GuCeM"
+              "ZJc4h"
             ],
             "typeId": "1be09af9-f089-4d6b-9f0b-48018b5f7393"
           }
@@ -30,12 +32,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "9c997f24-6859-4eff-a0ad-8fe63ffbb54e"
+        "x-ms-request-id": "b98ac990-fc3c-4ee3-9fbf-fed91ba1d491"
       },
       "ResponseBody": {
         "put": [
@@ -51,18 +53,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "50",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "60a39d59d02696b91eb065ade6f44cf8",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "2bf26355e0ee8cfbb87f67e826d1ac3d",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "get": {
           "timeSeriesIds": [
             [
-              "1uotX",
+              "URmuC",
               null,
-              "GuCeM"
+              "ZJc4h"
             ]
           ]
         }
@@ -72,12 +76,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "b611d418-1882-42a5-b116-36b25cb25a68"
+        "x-ms-request-id": "52e80ce9-9320-4a0b-86fa-d569d672a462"
       },
       "ResponseBody": {
         "get": [
@@ -85,9 +89,9 @@
             "instance": {
               "typeId": "1be09af9-f089-4d6b-9f0b-48018b5f7393",
               "timeSeriesId": [
-                "1uotX",
+                "URmuC",
                 null,
-                "GuCeM"
+                "ZJc4h"
               ]
             }
           }
@@ -102,18 +106,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "53",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "ed4fb0a65a46ec27280d17e6b5318f74",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "97d6975d8056a937130b3551dee1f573",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "delete": {
           "timeSeriesIds": [
             [
-              "1uotX",
+              "URmuC",
               null,
-              "GuCeM"
+              "ZJc4h"
             ]
           ]
         }
@@ -123,12 +129,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "9f8530ba-3bc2-49c8-afba-b253059a001b"
+        "x-ms-request-id": "a6e998c9-a469-4c45-a2b0-2081c061cec9"
       },
       "ResponseBody": {
         "delete": [
@@ -138,7 +144,7 @@
     }
   ],
   "Variables": {
-    "RandomSeed": "157988049",
+    "RandomSeed": "1795081276",
     "TIMESERIESINSIGHTS_URL": "fakeHost.api.wus2.timeseriesinsights.azure.com"
   }
 }
