+++ sdk/timeseriesinsights/Azure.IoT.TimeSeriesInsights/tests/SessionRecords/TimeSeriesIdTests/TimeSeriesId_CreateInstanceWith3KeysAsync.json
@@ -8,18 +8,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "97",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "740eb16e45a8acb13310b2f6ef70a52a",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "148a945b43aec96b7c6373cdb4926f7b",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "put": [
           {
             "timeSeriesId": [
-              "1uotX",
+              "xMhiS",
               null,
-              "GuCeM"
+              "yqYGZ"
             ],
             "typeId": "1be09af9-f089-4d6b-9f0b-48018b5f7393"
           }
@@ -30,12 +32,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "2cecb392-40e1-418e-ac9a-0c8a5dd727e0"
+        "x-ms-request-id": "b82a05ab-9f7f-43b9-9a57-314ba034bec8"
       },
       "ResponseBody": {
         "put": [
@@ -51,18 +53,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "50",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "60a39d59d02696b91eb065ade6f44cf8",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "1e243414fef4bdc09646a967d3f948ed",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "get": {
           "timeSeriesIds": [
             [
-              "1uotX",
+              "xMhiS",
               null,
-              "GuCeM"
+              "yqYGZ"
             ]
           ]
         }
@@ -72,12 +76,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "b91d133a-e7f4-433f-b642-a5e449359bae"
+        "x-ms-request-id": "38ebbb9a-f381-42f3-8b66-a0390a56aede"
       },
       "ResponseBody": {
         "get": [
@@ -85,9 +89,9 @@
             "instance": {
               "typeId": "1be09af9-f089-4d6b-9f0b-48018b5f7393",
               "timeSeriesId": [
-                "1uotX",
+                "xMhiS",
                 null,
-                "GuCeM"
+                "yqYGZ"
               ]
             }
           }
@@ -102,18 +106,20 @@
         "Authorization": "Sanitized",
         "Content-Length": "53",
         "Content-Type": "application/json",
-        "User-Agent": "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210401.1 (.NET Framework 4.8.4300.0; Microsoft Windows 10.0.19042 )",
-        "x-ms-client-request-id": "ed4fb0a65a46ec27280d17e6b5318f74",
-        "x-ms-client-session-id": "00000000-0000-0000-0000-000000000000",
+        "User-Agent": [
+          "azsdk-net-IoT.TimeSeriesInsights/1.0.0-alpha.20210405.1",
+          "(.NET 5.0.4; Microsoft Windows 10.0.19042)"
+        ],
+        "x-ms-client-request-id": "21bf3f17c5015ba25333425aec28905a",
         "x-ms-return-client-request-id": "true"
       },
       "RequestBody": {
         "delete": {
           "timeSeriesIds": [
             [
-              "1uotX",
+              "xMhiS",
               null,
-              "GuCeM"
+              "yqYGZ"
             ]
           ]
         }
@@ -123,12 +129,12 @@
         "Access-Control-Allow-Origin": "*",
         "Access-Control-Expose-Headers": "x-ms-request-id,x-ms-continuation",
         "Content-Type": "application/json",
-        "Date": "Fri, 02 Apr 2021 00:07:14 GMT",
+        "Date": "Mon, 05 Apr 2021 18:38:12 GMT",
         "Server": "Microsoft-HTTPAPI/2.0",
         "Strict-Transport-Security": "max-age=31536000; includeSubDomains",
         "Transfer-Encoding": "chunked",
         "X-Content-Type-Options": "nosniff",
-        "x-ms-request-id": "6af35859-574f-4ff4-a806-a52ffd3e4de1"
+        "x-ms-request-id": "6f56cb23-058a-4438-9e26-f6360fc4d93b"
       },
       "ResponseBody": {
         "delete": [
@@ -138,7 +144,7 @@
     }
   ],
   "Variables": {
-    "RandomSeed": "157988049",
+    "RandomSeed": "71271290",
     "TIMESERIESINSIGHTS_URL": "fakeHost.api.wus2.timeseriesinsights.azure.com"
   }
 }
